+++ taf/settings.py
@@ -29,6 +29,9 @@ last_validated_commit: dict = {}
 # determines if script files will be loaded from disk
 development_mode = False
 
+# determines if lifecycle handler scripts will be run
+run_scripts = False
+
 # The 'log.py' module manages TUF's logging system.  Users have the option to
 # enable/disable logging to a file via 'ENABLE_FILE_LOGGING', or
 # tuf.log.enable_file_logging() and tuf.log.disable_file_logging().